@JmPotato 2020-11-09
> How about local TSO off?

The QPS will both increase, however the performance gap ratio is still the same.