@ZenoTan 2020-11-09
If global TSO only, I think it should be the same as before. Why it is faster?