@JmPotato 2020-11-11
Here are some new results, PTAL @nolouch.

#### 1 client 100 concurrencies, 60s duration, requesting the Global TSO with Local TSO on.

This new version:

```shell
Total:
count:4036344, max:22, min:0, >1ms:3491867, >2ms:294228, >5ms:5392, >10ms:1370, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:4036344, >1ms:86.51%, >2ms:7.29%, >5ms:0.13%, >10ms:0.03%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:3772737, max:14, min:0, >1ms:3111424, >2ms:366387, >5ms:8273, >10ms:405, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:3772737, >1ms:82.47%, >2ms:9.71%, >5ms:0.22%, >10ms:0.01%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:3939770, max:17, min:0, >1ms:3278111, >2ms:353056, >5ms:5046, >10ms:796, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:3939770, >1ms:83.21%, >2ms:8.96%, >5ms:0.13%, >10ms:0.02%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Average: (4036344 + 3772737 + 3939770) * 1/3 * 1/60 = 65271 QPS
```

Previous bad performance version:

```shell
Total:
count:2576561, max:49, min:0, >1ms:1444939, >2ms:1023563, >5ms:50310, >10ms:9455, >30ms:699 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:2576561, >1ms:56.08%, >2ms:39.73%, >5ms:1.95%, >10ms:0.37%, >30ms:0.03% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:2725110, max:93, min:0, >1ms:1665302, >2ms:950362, >5ms:45013, >10ms:6950, >30ms:862 >50ms:315 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:2725110, >1ms:61.11%, >2ms:34.87%, >5ms:1.65%, >10ms:0.26%, >30ms:0.03% >50ms:0.01% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:2735030, max:29, min:0, >1ms:1634162, >2ms:993554, >5ms:48239, >10ms:4533, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:2735030, >1ms:59.75%, >2ms:36.33%, >5ms:1.76%, >10ms:0.17%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Average: (2576561 + 2725110 + 2735030) * 1/3 * 1/60 = 44648 QPS
```

Old good performance version:

```shell
Total:
count:4140666, max:18, min:0, >1ms:3708333, >2ms:140239, >5ms:1439, >10ms:1171, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:4140666, >1ms:89.56%, >2ms:3.39%, >5ms:0.03%, >10ms:0.03%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:4162000, max:25, min:0, >1ms:3779213, >2ms:123846, >5ms:2817, >10ms:300, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:4162000, >1ms:90.80%, >2ms:2.98%, >5ms:0.07%, >10ms:0.01%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:4082852, max:26, min:0, >1ms:3711833, >2ms:152472, >5ms:1405, >10ms:1639, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:4082852, >1ms:90.91%, >2ms:3.73%, >5ms:0.03%, >10ms:0.04%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%


Average: (4140666 + 4162000 + 4082852) * 1/3 * 1/60 = 68808 QPS
```

#### 1 client 100 concurrencies, 60s duration, requesting the Global TSO with Local TSO off.

This new version:

```shell
Total:
count:23839003, max:11, min:0, >1ms:36623, >2ms:20377, >5ms:2057, >10ms:21, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:23839003, >1ms:0.15%, >2ms:0.09%, >5ms:0.01%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:24152263, max:11, min:0, >1ms:26809, >2ms:14783, >5ms:776, >10ms:37, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:24152263, >1ms:0.11%, >2ms:0.06%, >5ms:0.00%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:23734834, max:14, min:0, >1ms:36441, >2ms:20559, >5ms:1846, >10ms:51, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:23734834, >1ms:0.15%, >2ms:0.09%, >5ms:0.01%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Average: (23839003 + 24152263 + 23734834) * 1/3 * 1/60 = 398478 QPS
```

Previous bad performance version:

```shell
Total:
count:12079401, max:13, min:0, >1ms:346864, >2ms:141912, >5ms:16352, >10ms:493, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:12079401, >1ms:2.87%, >2ms:1.17%, >5ms:0.14%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:11953634, max:19, min:0, >1ms:313129, >2ms:150919, >5ms:19237, >10ms:683, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:11953634, >1ms:2.62%, >2ms:1.26%, >5ms:0.16%, >10ms:0.01%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:12767788, max:11, min:0, >1ms:300035, >2ms:146473, >5ms:20260, >10ms:253, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:12767788, >1ms:2.35%, >2ms:1.15%, >5ms:0.16%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Average: (12079401 + 11953634 + 12767788) * 1/3 * 1/60 = 204449 QPS
```

Old good performance version:

```shell
Total:
count:21040046, max:39, min:0, >1ms:77983, >2ms:16566, >5ms:225, >10ms:0, >30ms:100 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:21040046, >1ms:0.37%, >2ms:0.08%, >5ms:0.00%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:20361055, max:39, min:0, >1ms:96215, >2ms:37119, >5ms:1187, >10ms:101, >30ms:100 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:20361055, >1ms:0.47%, >2ms:0.18%, >5ms:0.01%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Total:
count:20861289, max:7, min:0, >1ms:91560, >2ms:17761, >5ms:2, >10ms:0, >30ms:0 >50ms:0 >100ms:0 >200ms:0 >400ms:0 >800ms:0 >1s:0
count:20861289, >1ms:0.44%, >2ms:0.09%, >5ms:0.00%, >10ms:0.00%, >30ms:0.00% >50ms:0.00% >100ms:0.00% >200ms:0.00% >400ms:0.00% >800ms:0.00% >1s:0.00%

Average: (21040046 + 20361055 + 20861289) * 1/3 * 1/60 = 345902 QPS
```